Fix support for anonymous git repositories in builder
While the username and password for the git URL are, in theory,
optional, the image builder presently blows up if they are not defined,
owing to url_authentication referencing the missing dict keys.

Furthermore, the logic for determining whether to use authentication or
not did not take non-existence of keys into account, and was falsely
setting authentication to true when the keys were not defined in the
properties dict.