Execute Docker image build relative to cloned git repo
Dockerfiles expect to be running within the source tree, as they rely on
relative file paths for e.g. COPY operations. We use a two-step approach
here: (1) Prefer the Dockerfile in the top-level working directory, as
this is fetched independently; and (2) Provide the full path to the
cloned repository for the image build itself.

Fixes SODALITE-EU/image-builder#5