Support building image variants with layered base containers (#7)
* Execute Docker image build relative to cloned git repo

Dockerfiles expect to be running within the source tree, as they rely on
relative file paths for e.g. COPY operations. We use a two-step approach
here: (1) Prefer the Dockerfile in the top-level working directory, as
this is fetched independently; and (2) Provide the full path to the
cloned repository for the image build itself.

Fixes SODALITE-EU/image-builder#5

* Support building image variants with layered base containers

By default, the image builder only builds a single image and pushes it
out. This introduces an additional approach, in which multiple images
can be built from the Dockerfile by overloading the base container that
the image is built from.

As ARG is permitted to be inserted before FROM in the Dockerfile, we
achieve this by first mangling the Dockerfile in-place to permit base
image overloading through a specified BASE_IMAGE argument that the FROM
directive refers to. As an example:

	FROM golang as builder

becomes:

	ARG BASE_IMAGE=golang
	FROM ${BASE_IMAGE} as builder

This has been applied both to single-stage and multi-stage builders with
success.

The input file is responsible for defining all of the different image
variant/base combinations that it wants to generate. While in practice
this will be used for layering services on top of different accelerator
runtimes, a synthetic example providing tagged images for different
versions of the language runtime is as follows:

	target:
	  registry_ip: localhost:5000
	  images:
	  - image: adaptant/edgetpu-exporter
	    tag: latest

	  - image: adaptant/edgetpu-exporter
	    tag: go1.13
	    base: golang:1.13

The image name can be, but must not, be the same for each variant. The
omission of the base property will default to the existing base image
defined within the Dockerfile.

Co-authored-by: mihaTrajbaric <26118837+mihaTrajbaric@users.noreply.github.com>