Allow subdir relative build context for Git repos
This adds an optional 'subdir' attribute to the git build context
which allows for the build contet to be set relative to a path
within the cloned repository.

This can be useful in cases where a given repository contains
multiple Dockerfiles in different subdirectories, and the build must
be run relative to the subdirectory itself.